Move PLP commands and parsers in the merlin's connector
