+++ web/app/containers/product-list/container.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import * as commands from '../../integration-manager/commands'
+import {fetchProductListData} from '../../integration-manager/categories/commands'
 import {isRunningInAstro} from '../../utils/astro-integration'
 import ProductListHeader from './partials/product-list-header'
 import ProductListContents from './partials/product-list-contents'
@@ -17,7 +17,7 @@ const ProductList = () => {
 }
 
 ProductList.fetcher = (url, dispatch) => {
-    dispatch(commands.fetchProductListData(url))
+    dispatch(fetchProductListData(url))
 }
 
 export default ProductList
