+++ web/app/integration-manager/_merlins-connector/app/commands.js
@@ -1,8 +1,8 @@
 import {jqueryResponse} from 'progressive-web-sdk/dist/jquery-response'
 import {makeRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
-import {parseNavigation} from './navigation/parser'
+import {parseNavigation} from '../navigation/parser'
 
-import {receiveNavigationData} from '../responses'
+import {receiveNavigationData} from '../../responses'
 
 export const fetchPageData = (url) => (dispatch) => {
     return makeRequest(url)
