+++ web/app/integration-manager/_merlins-connector/products/commands.js
@@ -3,7 +3,7 @@ import {receiveFormInfo} from '../../actions'
 
 import {fetchPageData} from '../app/commands'
 
-import {receiveProductDetailsProductData, receiveProductDetailsUIData} from '../../product-details/responses'
+import {receiveProductDetailsProductData, receiveProductDetailsUIData} from '../../products/responses'
 import {productDetailsParser, productDetailsUIParser, pdpAddToCartFormParser} from './parsers'
 
 export const fetchPdpData = (url) => (dispatch) => {
