+++ web/app/integration-manager/_merlins-connector/products/product-details-example.html
