+++ web/app/integration-manager/categories/commands.js
@@ -1,0 +1,7 @@
+let connector = {}
+
+export const register = (commands) => {
+    connector = commands
+}
+
+export const fetchProductListData = (...args) => connector.fetchProductListData(...args)
