+++ web/app/integration-manager/categories/responses.js
@@ -1,0 +1,3 @@
+import {createAction} from '../../utils/utils'
+
+export const receiveCategory = createAction('Receive Category data')
