+++ web/app/integration-manager/products/commands.js
