+++ web/app/store/categories/reducer.js
@@ -1,7 +1,7 @@
 import Immutable from 'immutable'
 import {handleActions} from 'redux-actions'
 import {mergePayloadForActions} from '../../utils/reducer-utils'
-import {receiveCategory} from '../../integration-manager/responses'
+import {receiveCategory} from '../../integration-manager/categories/responses'
 
 const initialState = Immutable.Map()
